log post and put calls properly
